move ggml_qnn_op_config to deparated file
